+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.1-dev
+
+- `DefaultJsonWebKeySetLoader`: if possible, use HTTP headers to determine
+  cache expiration. 
+
 ## 0.3.0
 
 - Migrate null safety